requirements.txt
aioredis==1.3.1
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.21.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
ConfigArgParse==1.2.3
cryptography==3.2.1
databases==0.4.1
ecdsa==0.14.1
fastapi==0.63.0
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.0.0
h11==0.12.0
hiredis==1.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.1.0
locust==1.4.1
MarkupSafe==1.1.1
msgpack==1.0.2
packaging==20.8
passlib==1.7.4
pluggy==0.13.1
psutil==5.8.0
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.1
python-decouple==3.4
python-jose==3.2.0
python-multipart==0.0.5
pyzmq==21.0.1
requests==2.25.1
rsa==4.7
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
toml==0.10.2
urllib3==1.26.2
uvicorn==0.13.3
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
